Implement support for SLO uploads to Google.
First pass at SLO uploads to Google. Since GCS allows up to 5TB
objects, we do not use the compose mechanism. Each SLO is uploaded as
a single object. For that we have an SLOFileWrapper class that takes a
manifest and produces a single object stream (by requesting the
appropriate parts of the manifest).

Just like AWS S3 uploads, we do not yet support range manifests here.