Add an abstraction for the sync providers.
As we add support for Swift, we need to add an abstraction for the
sync mechanisms. The required interface for each of the sync providers
is:
* upload_object(name, storage_policy)
* delete_object(name)

One additional internal method must be implemented by every
subclass: _get_client_factory(). This method returns a factory that
can produce the clients for the provider. In this way, the
HttpClientPool can be repatriated into the base provider class
(BaseSync) and all classes that implement the functionality derive
from it.

This patch moves the existing sync to S3 to this model via the SyncS3
class.