Implement Sync to Swift.
Uses python-swiftclient to sync data from a Swift cluster to another
Swift cluster. Does not support DLO manifests. SLOs are propagated to
the remote end. On deletion, SLO segments are removed as well. If
there are multiple manifests that reference the same object, the parts
of the object will be removed when the first manifest referring to it
is removed. Ranges are not supported in the SLOs propagated to the
remote cluster.