Add copy_after, retain_local, propagate_delete.
Adds three new configuration options for each container. This allows
for expressing more complex policies, such as archival.

For the case of not retaining a local copy and copying after a set
amount of time, we leverage the X-Timestamp header with the DELETE
verb to ensure that we will not overwrite an object that may have been
PUT while we were copying data into the public cloud.