Handle 409 Conflict on DELETE with X-Timestamp.
When we attempt to DELETE an object after copy (if retain_local is set
to False), we should be prepared to handle a 409 Conflict error. This
occurs if the object was actually overwritten while we were copying
and we attempt to issue a DELETE with an X-Timestamp header set in the
past.