Establish Swift connection after pool.get_client.
We should only establish a Swift connection after we are ready to be
transmitting data. Currently, we may connect using InternalClient and
then wait > 60 seconds for the remote cloud client (in the acquire
call). In the meantime, Swift will close our connection with an
exception.

The patch moves the Swift connection establishment inside the "with"
clause used to obtain the remote client.