Use multiple simultaneous connections.
Previously, we were blocking when attempting to acquire a connection.
This resulted in using only one connection at a time (once we set the
number of connections per Boto object to 1).

However, that behavior was masking a bug where we were not holding the
semaphore representing the total connection count until after we were
done using the client. The patch fixes both of these behaviors.