Hacky support for HTTP/HTTPS proxies.
boto lacks the plumbing to set proxies on a per-connection basis, so
we just set it globally.

It's done in main() since it doesn't really belong anywhere
else. ContainerCrawler doesn't know anything about networks or HTTP,
so it can't go there. SyncS3 and SyncSwift shouldn't do it because
it's a global setting and there's a whole bunch of sync objects. That
leaves the bit of plumbing where they intersect, hence this diff.