Make swiftstack_auth optional.
We patch the logger used for middleware to work around a Swift bug
that results in transaction IDs not being printed at INFO level for
middleware. This bug has since been fixed and we can make this patch
optional. This allows using cloud sync outside of SwiftStack and in
the future we probably want to remove this entirely.