Implement SLO restore on a GET.
Implements fetching large objects out of the blob store when a GET
request is issued (and the "restore_object" option is set). This
change affects the SLO upload, as in addition to the SLO itself, we
now also upload the manifest.

If the manifest is not found -- meaning the object was uploaded prior
to this change -- it cannot be restored, as we don't know how it is
segmented and where the segments should be placed.

This means that when deleting objects, we also remove the accompanying
manifest. This is done unconditionally, as there is no benefit to
issuing a prior HEAD.

On restore, we attempt to fetch the manifest first, as we cannot issue
the object GET -- since we never read the full object body (it is
returned to wsgi, wrapped in the "Swift[Slo]PutWrapper"). This means
that a GET, with the restore_object option set, results in two
requests -- GET for the manifest and the object.