Fix SLO header validation on uploads.
When uploading SLOs, we need to validate that the SLO header is
present on the S3 object. Previously, we would possibly re-upload
objects, as the check did not take into account that the header would
appear in the S3 keys, but would not appear in the Swift user
metadata tags.