Add POST support for archived objects.
Propagate the POST requests to the archive store if the object is not
found in Swift. We have to issue a HEAD request before propagating the
HEAD in the case of S3/Google, as otherwise we may overwrite headers
that cloud sync tracks (e.g. the SLO ETag field for GCS or the
X-Static-Large-Object field). Similarly, we have to retrieve the
Content-Type and pass it on the server-side copy (PUT), as otherwise
the storage systems will use the default Content-Type (i.e.
application/octet-stream for S3).