Remove a dependence on container_crawler.
The sync_swift module is used in the shunt code, which is loaded by
the Swift proxy server. Adding a dependecy for the RetryError class
means that the container_crawler library must be available, as well.

In some cases, it may not be and adding the dependency just for the
RetryError is a little silly. This patch changes the exception to a
RuntimeError (until we have a better way to do this).