Add get_object and head_object to the BaseSync.
We need to be able to issue GET requests to the remote containers not
just as part of shunting requests, but also when migrating objects
back into Swift. This patchs expands the existing API, such that the
migrator could LIST and GET all of the objects from the remote store
and stream them into Swift.