Return connections only after consuming the body.
When handling GET requests through the shunt, we should make sure the
body is consumed before returning the connection to the pool. This
means that we cannot use the context manager without turning the
get_object() function into a generator. The patch adds a wrapper
around the body that returns the connection to the pool after the body
is fully consumed (or the object is destroyed).