Add a handle_proxyfs toggle
When using cloud sync with proxyfs, we have to sync the log segment data
(because that's what will actually be listed in the containers we're
crawling), which means the cloud shunt should go *after* pfs_middleware
in the pipeline.

However, we also have to be *before* SLO in the pipeline to be able to
nicely translate between SLOs and Multipart Uploads. Otherwise, SLO will
try to read the large object as though it was a manifest (!) since we
add the X-Static-Large-Object header to the shunted response.

Fortunately, we'll never have SLOs to deal with in bimodal accounts, so
we can resolve this by having the shunt in the pipeline twice -- first
to handle normal swift accounts (ahead of SLO), then to handle bimodal
accounts (after pfs_middleware), with flags to ensure that any given
request will only be processed once.