Keep migrator scan and moved counts for last run in status file
In addition to the scan & count values for the current cycle tracked in
the migrator status file, when a cycle finishes the current values will
be preserved in keys prefixed with "last_".  These values can be
collected by external monitoring tools to provide a measure of progress
for each migration.